refactor: [#36] AddButton.tsx의 공통 컴포넌트화
